chore: move readpipedatatask logic to waylandcopyclient
move the logic to waylandcopyclient, seems it will reduced when the
signal cannot received by task

Log:
